requirements.txt
flask
playwright
openai
python-dotenv
markdown
